fix: migrate credit system from legacy credits column to subscription_allowance + purchased_credits
- Remove int() coercion on credits in /user/profile (was truncating float balances)
- Update /user/balance to compute total from subscription_allowance + purchased_credits
- Update add_credits and adjust_credits endpoints to write to purchased_credits column
- Update DB queries across users, coupons, credit_transactions to use new schema fields
- Update auth, schemas, and security deps to reflect tiered credit fields
- Add migration to drop legacy credits column
- Update CLAUDE.md and AGENTS.md

+++ src/schemas/auth.py
@@ -146,8 +146,8 @@ class PrivyAuthResponse(BaseModel):
     tier_display_name: str | None = None
     trial_expires_at: str | None = None
     subscription_end_date: int | None = None
-    # Tiered credit fields (in cents for frontend consistency)
-    subscription_allowance: int | None = None  # Monthly subscription allowance
-    purchased_credits: int | None = None  # One-time purchased credits
-    total_credits: int | None = None  # Sum of subscription_allowance + purchased_credits
+    # Tiered credit fields — 1 credit = $1, stored as float
+    subscription_allowance: float | None = None  # Monthly subscription allowance
+    purchased_credits: float | None = None  # One-time purchased credits
+    total_credits: float | None = None  # Sum of subscription_allowance + purchased_credits
     allowance_reset_date: str | None = None  # When allowance was last reset
